src/test/java/net/ryandoyle/libjnagios/builder/HtmlHostBuilderTest.java
package net.ryandoyle.libjnagios.builder;

import net.ryandoyle.libjnagios.domain.Host;
import org.jsoup.Jsoup;
import org.jsoup.nodes.Document;
import org.junit.Before;
import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;
import static org.mockito.MockitoAnnotations.initMocks;

public class HtmlHostBuilderTest extends BaseHtmlTestFactory {

    Host host;

    @Before
    public void setup(){
        initMocks(this);
        Document document = Jsoup.parse(html);
        host = new HtmlHostBuilder(document).build();
    }

    @Test
    public void itShouldCreateAHostFromAValidHtmlString(){
        assertThat(host.getName(), is("localhost"));
    }

    @Test
    public void theHostShouldBePopulatedWithServices(){
        assertThat(host.getServices().isEmpty(), is(false));
    }

}